+++ actix-server/src/lib.rs
@@ -14,8 +14,8 @@ mod test_server;
 mod waker_queue;
 mod worker;
 
-pub use self::builder::ServerBuilder;
-pub use self::server::Server;
+pub use self::builder::{Server, ServerBuilder};
+pub use self::server::{ServerHandle};
 pub use self::service::ServiceFactory;
 pub use self::test_server::TestServer;